Remove one-to-many-dms feature flag
