+++ packages/common/src/hooks/chats/index.ts
@@ -1,7 +1,6 @@
 export * from './types'
 export * from './useAudienceUsers'
 export * from './useCanSendMessage'
-export * from './useSetInboxPermissionsLegacy'
 export * from './useSetInboxPermissions'
 export * from './useTrackPlayer'
 export * from './useAudiusLinkResolver'
