+++ packages/common/src/services/remote-config/feature-flags.ts
@@ -17,7 +17,6 @@ export enum FeatureFlags {
   COMMENTS_ENABLED = 'comments_enabled',
   COMMENT_POSTING_ENABLED = 'comment_posting_enabled',
   PAID_SCHEDULED = 'paid_scheduled',
-  ONE_TO_MANY_DMS = 'one_to_many_dms',
   NETWORK_CUT_ENABLED = 'network_cut_enabled',
   SOCIAL_SIGNUP = 'social_signup',
   RIGHTS_AND_COVERS = 'rights_and_covers',
@@ -58,7 +57,6 @@ export const flagDefaults: FlagDefaults = {
   [FeatureFlags.COMMENTS_ENABLED]: false,
   [FeatureFlags.COMMENT_POSTING_ENABLED]: false,
   [FeatureFlags.GUEST_CHECKOUT]: false,
-  [FeatureFlags.ONE_TO_MANY_DMS]: false,
   [FeatureFlags.NETWORK_CUT_ENABLED]: false,
   [FeatureFlags.SOCIAL_SIGNUP]: false,
   [FeatureFlags.RIGHTS_AND_COVERS]: false,
