--- packages/web/src/components/inbox-settings-modal/InboxSettingsModalLegacy.module.css
@@ -1,14 +1,0 @@
-.modalBody {
-  max-width: 640px;
-}
-
-.modalContent > div {
-  display: flex;
-  flex-direction: column;
-}
-
-.error {
-  margin: var(--harmony-unit-6) auto 0;
-  font-weight: var(--harmony-font-demi-bold);
-  color: var(--harmony-red);
-}
